+++ src/main/java/org/mule/module/apikit/Configuration.java
@@ -15,15 +15,15 @@ import org.mule.apikit.ApiType;
 import org.mule.module.apikit.api.RamlHandler;
 import org.mule.module.apikit.api.config.ConsoleConfig;
 import org.mule.module.apikit.api.config.ValidationConfig;
+import org.mule.module.apikit.api.deserializing.AttributesDeserializingStrategy;
+import org.mule.module.apikit.api.deserializing.AttributesDeserializingStrategyIdentifier;
 import org.mule.module.apikit.api.exception.ApikitRuntimeException;
-import org.mule.module.apikit.api.parsing.AttributesDeserializingStrategy;
-import org.mule.module.apikit.api.parsing.AttributesDeserializingStrategyIdentifier;
 import org.mule.module.apikit.api.spi.RouterService;
 import org.mule.module.apikit.api.uri.URIPattern;
 import org.mule.module.apikit.api.uri.URIResolver;
 import org.mule.module.apikit.api.validation.ApiKitJsonSchema;
-import org.mule.module.apikit.parsing.AttributesDeserializingStrategies;
-import org.mule.module.apikit.parsing.NoneAttributeDeserializingStrategy;
+import org.mule.module.apikit.deserializing.AttributesDeserializingStrategies;
+import org.mule.module.apikit.deserializing.NoneAttributeDeserializingStrategy;
 import org.mule.module.apikit.validation.body.schema.v1.cache.JsonSchemaCacheLoader;
 import org.mule.module.apikit.validation.body.schema.v1.cache.XmlSchemaCacheLoader;
 import org.mule.runtime.api.component.location.ConfigurationComponentLocator;
