+++ src/main/java/org/mule/module/apikit/api/deserializing/AttributesDeserializingStrategy.java
@@ -4,12 +4,13 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.module.apikit.api.parsing;
+package org.mule.module.apikit.api.deserializing;
 
 /**
- * Set of identifiers for attribute parsing strategies.
+ * Attribute deserializing strategies set by configuration will implement this interface.
  */
-public enum AttributesDeserializingStrategyIdentifier {
-  NONE_DESERIALIZING_STRATEGY,
-  ARRAY_HEADER_DESERIALIZING_STRATEGY
+public interface AttributesDeserializingStrategy {
+
+  AttributesDeserializingStrategyIdentifier getStrategyIdentifier();
+
 }
