+++ src/main/java/org/mule/module/apikit/deserializing/ArrayHeaderDelimiter.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.module.apikit.parsing;
+package org.mule.module.apikit.deserializing;
 
 /**
  * Set of values that are allowed as delimiters for HTTP array header values.
