+++ src/main/java/org/mule/module/apikit/deserializing/ArrayHeaderDelimiterTypeConverter.java
@@ -4,12 +4,12 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.module.apikit.parsing;
+package org.mule.module.apikit.deserializing;
 
 import org.mule.runtime.dsl.api.component.TypeConverter;
 
-import static org.mule.module.apikit.parsing.ArrayHeaderDelimiter.COMMA;
-import static org.mule.module.apikit.parsing.ArrayHeaderDelimiter.NONE;
+import static org.mule.module.apikit.deserializing.ArrayHeaderDelimiter.COMMA;
+import static org.mule.module.apikit.deserializing.ArrayHeaderDelimiter.NONE;
 
 /**
  * Maps the String delimiter set in configuration with the corresponding enum value
