+++ src/test/java/org/mule/module/apikit/ConfigurationTest.java
@@ -9,16 +9,16 @@ package org.mule.module.apikit;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
-import org.mule.module.apikit.api.parsing.AttributesDeserializingStrategy;
-import org.mule.module.apikit.parsing.ArrayHeaderDeserializingStrategy;
-import org.mule.module.apikit.parsing.AttributesDeserializingStrategies;
-import org.mule.module.apikit.parsing.NoneAttributeDeserializingStrategy;
+import org.mule.module.apikit.api.deserializing.AttributesDeserializingStrategy;
+import org.mule.module.apikit.deserializing.ArrayHeaderDeserializingStrategy;
+import org.mule.module.apikit.deserializing.AttributesDeserializingStrategies;
+import org.mule.module.apikit.deserializing.NoneAttributeDeserializingStrategy;
 
 import static java.util.Arrays.asList;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
-import static org.mule.module.apikit.api.parsing.AttributesDeserializingStrategyIdentifier.ARRAY_HEADER_DESERIALIZING_STRATEGY;
-import static org.mule.module.apikit.api.parsing.AttributesDeserializingStrategyIdentifier.NONE_DESERIALIZING_STRATEGY;
+import static org.mule.module.apikit.api.deserializing.AttributesDeserializingStrategyIdentifier.ARRAY_HEADER_DESERIALIZING_STRATEGY;
+import static org.mule.module.apikit.api.deserializing.AttributesDeserializingStrategyIdentifier.NONE_DESERIALIZING_STRATEGY;
 
 public class ConfigurationTest {
 
@@ -26,7 +26,7 @@ public class ConfigurationTest {
   public ExpectedException expectedException = ExpectedException.none();
 
   @Test
-  public void getAttributeParsingStrategyByIdentifier() {
+  public void getAttributeDeserializingStrategyByIdentifier() {
     AttributesDeserializingStrategies strategies = new AttributesDeserializingStrategies();
     strategies.setAttributesDeserializingStrategies(asList(new ArrayHeaderDeserializingStrategy(),
                                                            new NoneAttributeDeserializingStrategy()));
@@ -45,19 +45,19 @@ public class ConfigurationTest {
   public void getDummyStrategyIfNoStrategies() {
     Configuration configuration = new Configuration();
     configuration.setAttributesDeserializingStrategies(new AttributesDeserializingStrategies());
-    AttributesDeserializingStrategy parsingStrategy =
+    AttributesDeserializingStrategy deserializingStrategy =
         configuration.getAttributesDeserializingStrategy(ARRAY_HEADER_DESERIALIZING_STRATEGY);
-    assertNotNull(parsingStrategy);
-    assertTrue(NoneAttributeDeserializingStrategy.class.isInstance(parsingStrategy));
+    assertNotNull(deserializingStrategy);
+    assertTrue(NoneAttributeDeserializingStrategy.class.isInstance(deserializingStrategy));
   }
 
   @Test
   public void getDummyStrategyIfNullStrategies() {
     Configuration configuration = new Configuration();
-    AttributesDeserializingStrategy parsingStrategy =
+    AttributesDeserializingStrategy deserializingStrategy =
         configuration.getAttributesDeserializingStrategy(ARRAY_HEADER_DESERIALIZING_STRATEGY);
-    assertNotNull(parsingStrategy);
-    assertTrue(NoneAttributeDeserializingStrategy.class.isInstance(parsingStrategy));
+    assertNotNull(deserializingStrategy);
+    assertTrue(NoneAttributeDeserializingStrategy.class.isInstance(deserializingStrategy));
   }
 
   @Test
