+++ src/test/java/org/mule/module/apikit/deserializing/ArrayHeaderDelimiterTypeConverterTest.java
@@ -4,13 +4,13 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.module.apikit.parsing;
+package org.mule.module.apikit.deserializing;
 
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
-import static org.mule.module.apikit.parsing.ArrayHeaderDelimiter.COMMA;
-import static org.mule.module.apikit.parsing.ArrayHeaderDelimiter.NONE;
+import static org.mule.module.apikit.deserializing.ArrayHeaderDelimiter.COMMA;
+import static org.mule.module.apikit.deserializing.ArrayHeaderDelimiter.NONE;
 
 public class ArrayHeaderDelimiterTypeConverterTest {
