+++ src/test/java/org/mule/module/apikit/deserializing/DummyAttributeDeserializerTest.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.module.apikit.parsing;
+package org.mule.module.apikit.deserializing;
 
 import org.junit.Before;
 import org.junit.Test;
