+++ src/test/java/org/mule/module/apikit/validation/AbstractRequestValidatorTestCase.java
@@ -14,9 +14,9 @@ import org.mule.extension.http.api.HttpRequestAttributes;
 import org.mule.extension.http.api.HttpRequestAttributesBuilder;
 import org.mule.module.apikit.api.RoutingTable;
 import org.mule.module.apikit.api.config.ValidationConfig;
+import org.mule.module.apikit.api.deserializing.AttributesDeserializingStrategy;
+import org.mule.module.apikit.api.deserializing.AttributesDeserializingStrategyIdentifier;
 import org.mule.module.apikit.api.exception.MuleRestException;
-import org.mule.module.apikit.api.parsing.AttributesDeserializingStrategy;
-import org.mule.module.apikit.api.parsing.AttributesDeserializingStrategyIdentifier;
 import org.mule.module.apikit.api.uri.URIPattern;
 import org.mule.module.apikit.api.uri.URIResolver;
 import org.mule.module.apikit.api.uri.URIResolver.MatchRule;
